Fix secret.env permissions

+++ init-config.sh
@@ -16,12 +16,10 @@ read -p "Create secret.env ? (y/N) " -r
 if [[ $REPLY =~ ^[Yy]$ ]]; then
   SECRET_KEY=$(LC_ALL=C tr -dc '[:alnum:]' < /dev/urandom | head -c40)
   SECRET_KEY_BASE64=$(echo $SECRET_KEY | base64)
-  cp secret.env secret.env.bak
+  [ -f secret.env ] && cp secret.env secret.env.bak
   echo "SECRET_KEY=$SECRET_KEY" > secret.env
   echo "SECRET_KEY_BASE64=$SECRET_KEY_BASE64" >> secret.env
   echo "DJANGO_SUPERUSER_PASSWORD=$(LC_ALL=C tr -dc '[:alnum:]' < /dev/urandom | head -c15)" >> secret.env
-
-  chown $OWNER secret.env # change ownership of file created
 fi
 
 echo -e "\n### Creating RSA key pair for JWT (conf/keys/jwt.priv.pem). This will replace old keys (if exist; backup will be in data/keys/jwt.priv.pem.bak)."
